# find-my-mua/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Find My MUA API",
        "description": "Dokumentasi lengkap API untuk aplikasi Find My MUA",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/register/customer": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register akun customer",
                "operationId": "562ed09c77f8412f14906aefa38ac068",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "anna@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "secret"
                                    },
                                    "skin_tone": {
                                        "type": "string",
                                        "example": "sawo matang"
                                    },
                                    "skin_type": {
                                        "type": "string",
                                        "example": "berminyak"
                                    },
                                    "makeup_preferences": {
                                        "type": "string",
                                        "example": "natural, matte"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/mua/{id}/availability": {
            "get": {
                "tags": [
                    "Availability"
                ],
                "summary": "Lihat slot waktu yang masih tersedia untuk MUA",
                "operationId": "0d94cecaa6534e79774a46d02dda5cf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot tersedia dikembalikan"
                    }
                }
            }
        },
...
        "/api/customer/bookings": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Lihat semua booking milik customer",
                "operationId": "a3442a0a9acd411936ad4f5673735660",
                "responses": {
                    "200": {
                        "description": "Daftar booking"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Buat booking baru",
                "operationId": "753e234378345871167a3a04b155d307",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mua_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "14:00"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "transfer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/customer/reviews": {
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Customer menulis review setelah booking selesai",
                "operationId": "ee01d8b044bd61c97109d04b6078e79b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "booking_id": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Hasil makeup bagus dan awet!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review disimpan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/mua/bookings": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Lihat semua booking untuk MUA",
                "operationId": "3af2ad63f8140f7ab534cafaf1ff17a5",
                "responses": {
                    "200": {
                        "description": "List booking untuk MUA"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mua/bookings/{id}/status": {
            "put": {
                "tags": [
                    "Booking"
                ],
                "summary": "Ubah status booking oleh MUA",
                "operationId": "31daa28d293e50e38ba7bb2a75828b10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "example": "paid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status booking diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}